requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
autograd==1.8.0
autograd-gamma==0.5.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmdstanpy==1.2.5
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
formulaic==1.2.1
gitdb==4.0.12
GitPython==3.1.45
holidays==0.82
idna==3.11
importlib_resources==6.5.2
interface-meta==1.3.0
invoke==2.2.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lifelines==0.30.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
narwhals==2.8.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
prophet==1.1.7
protobuf==6.33.0
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sshtunnel==0.4.0
stack-data==0.6.3
stanio==0.5.1
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
wrapt==1.17.3
xgboost==3.0.5
